implemented the 'Router'
With the addition of the 'Router' I don't need to have so many values
inside my store, because mosts thing will be managed using the router,
so the code in 'App' is now cleaner. After finnishing the implementation,
I will have less components in the App  and less global values inside store.